+++ cpp/tests/layout/trust_worthiness.h
@@ -16,6 +16,10 @@
 
 #include "knn.h"
 
+#include <algorithm>
+#include <iostream>
+#include <vector>
+
 double euclidian_dist(const std::vector<int>& x, const std::vector<int>& y)
 {
   double total = 0;