+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## New Features
 - PR #937 Add wrapper for gunrock HITS algorithm
 - PR #939 Updated Notebooks to include new features and benchmarks
-- PR #944 opg pagerank dask
+- PR #944 OPG pagerank (dask)
+- PR #947 OPG pagerank (CUDA)
 
 ## Improvements
 - PR #898 Add Edge Betweenness Centrality, and endpoints to BC
@@ -19,6 +20,7 @@
 - PR #946 Install meta packages for dependencies
 - PR #952 Updated get_test_data.sh to also (optionally) download and install datasets for benchmark runs
 - PR #953 fix setting RAFT_DIR from the RAFT_PATH env var
+- PR #954 Update cuGraph error handling to use RAFT
 
 ## Bug Fixes
 - PR #936 Update Force Atlas 2 doc and wrapper